# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Recetas DWES",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/recetas/{receta}/ingredientes": {
            "get": {
                "tags": [
                    "Ingredientes"
                ],
                "summary": "Listar ingredientes de una receta",
                "operationId": "26cb65af165bcb7cd201915fcb8aecf7",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ingredientes"
                ],
                "summary": "Agregar un ingrediente a una receta",
                "operationId": "48b90f075f51df7432c51970cb3de40f",
                "responses": {
                    "201": {
                        "description": "Ingrediente creado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/ingredientes/{ingrediente}": {
            "get": {
                "tags": [
                    "Ingredientes"
                ],
                "summary": "Ver un ingrediente específico",
                "operationId": "9641bc78a22bc6ac50a9777bb90ca23d",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Ingredientes"
                ],
                "summary": "Actualizar un ingrediente",
                "operationId": "93bd749a16de4bce1cb009d5425ff1a0",
                "responses": {
                    "200": {
                        "description": "Ingrediente actualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ingredientes"
                ],
                "summary": "Eliminar un ingrediente",
                "operationId": "e23088e34bc8e43c821256e2c7ae5ec7",
                "responses": {
                    "200": {
                        "description": "Ingrediente eliminado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/recetas/{receta}/like": {
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Dar/quitar like (toggle)",
                "description": "Alterna el estado de like: si existe lo elimina, si no existe lo crea",
                "operationId": "eddfc38be8bef694304716bc3b01b124",
                "parameters": [
                    {
                        "name": "receta",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Like añadido correctamente"
                    },
                    "200": {
                        "description": "Like eliminado correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Receta no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/recetas/{receta}/likes/count": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Obtener contador de likes",
                "description": "Retorna el número total de likes de una receta",
                "operationId": "e8ced5a8993bdbf028fa1220f635bbf3",
                "parameters": [
                    {
                        "name": "receta",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contador de likes obtenido"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/recetas/{receta}/likes": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Listar usuarios que dieron like",
                "description": "Obtiene la lista de usuarios que dieron like a una receta",
                "operationId": "55b3d9c3ce7d1fee23c4e9d0ccd08a80",
                "parameters": [
                    {
                        "name": "receta",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuarios que dieron like"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}